requirements.txt
# ----------------------------------------------------------------------
# 1. CORE API FRAMEWORK
# FastAPI is the high-performance framework for the backend API server.
# Uvicorn is the ASGI server that runs the application.
# ----------------------------------------------------------------------
fastapi>=0.104.1
uvicorn[standard]>=0.23.2

# ----------------------------------------------------------------------
# 2. RAG ORCHESTRATION FRAMEWORKS
# We will use LangChain (or LlamaIndex) for document loading, chunking,
# and coordinating the RAG process.
# ----------------------------------------------------------------------
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-huggingface>=0.0.7 # For local HuggingFace LLM wrappers
langchain-openai>=0.1.0  # For OpenAI API support (backup option)

# ----------------------------------------------------------------------
# 3. EMBEDDING MODEL & VECTOR STORE
# sentence-transformers for local/HuggingFace embeddings.
# chromadb for the lightweight, local vector database.
# langchain-chroma provides the LangChain integration with ChromaDB.
# ----------------------------------------------------------------------
sentence-transformers>=2.2.2 # For local embeddings (e.g., all-MiniLM-L6-v2)
chromadb>=0.4.15
langchain-chroma>=0.1.0
torch>=2.1.0
transformers>=4.38.0
accelerate>=0.25.0
sentencepiece>=0.1.99
pydantic>=2.5.2 # Used by FastAPI, LangChain, and ChromaDB

# ----------------------------------------------------------------------
# 4. UTILITIES & DATA HANDLING
# python-dotenv for managing environment variables (API keys).
# pydantic is also used here for data validation across the system.
# The 'tiktoken' library is for token counting (used by OpenAI models).
# ----------------------------------------------------------------------
python-dotenv>=1.0.0

# ----------------------------------------------------------------------
# 5. CODE PROCESSING (CRITICAL FOR CODE ASSISTANT)
# Pygment is a code syntax highlighter and parser which helps with
# intelligent code chunking for many languages.
# ----------------------------------------------------------------------
Pygments>=2.17.2

# ----------------------------------------------------------------------
# 6. AWS INTEGRATION FOR REMOTE LLM INFERENCE
# ----------------------------------------------------------------------
boto3>=1.28.0

# ----------------------------------------------------------------------
# 7. GOOGLE CLOUD VERTEX AI INTEGRATION
# ----------------------------------------------------------------------
google-cloud-aiplatform>=1.38.0
httpx>=0.25.0  # For graph API calls